feat: video display on image gallery overlay
