+++ examples/ExpoMessaging/package.json
@@ -12,6 +12,7 @@
     "@react-navigation/native": "5.8.10",
     "@react-navigation/stack": "5.12.8",
     "expo": "^44.0.0",
+    "expo-av": "~10.2.0",
     "expo-blur": "~11.0.0",
     "expo-document-picker": "~10.1.3",
     "expo-file-system": "~13.1.4",
@@ -20,6 +21,8 @@
     "expo-image-picker": "~12.0.1",
     "expo-media-library": "~14.0.0",
     "expo-sharing": "~10.1.0",
+    "expo-splash-screen": "~0.14.1",
+    "expo-status-bar": "~1.2.0",
     "react": "17.0.1",
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
@@ -28,11 +31,9 @@
     "react-native-safe-area-context": "3.3.2",
     "react-native-screens": "~3.10.1",
     "react-native-svg": "12.1.1",
+    "react-native-web": "0.17.1",
     "stream-chat-expo": "link:../../package/expo-package",
-    "stream-chat-react-native-core": "link:../../package",
-    "expo-splash-screen": "~0.14.1",
-    "expo-status-bar": "~1.2.0",
-    "react-native-web": "0.17.1"
+    "stream-chat-react-native-core": "link:../../package"
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
